+++ src/App.tsx
@@ -1,19 +1,15 @@
 import { useEffect, useState, useMemo } from 'react'
 import SpotifyWebApi from 'spotify-web-api-js'
 import Login from './components/Login'
-import MainPage from './components/pages/MainPage/MainPage'
+import MainPage from './pages/MainPage/MainPage'
 import { getTokenFromUrl } from './helper/spotify'
 import { Route, Switch } from 'react-router-dom'
-import RecommendationPage from './components/pages/RecommendationPage/RecommendationPage'
+import RecommendationPage from './pages/RecommendationPage/RecommendationPage'
 
 function App(): JSX.Element {
   const [tracks, setTracks] = useState<SpotifyApi.TrackObjectSimplified[]>()
 
-  useEffect(() => {
-    tracks?.forEach(track => console.log(track.name))
-  }, [tracks])
-
-  const token = useMemo(getTokenFromUrl, [])
+  const token: string | undefined = useMemo(getTokenFromUrl, [])
 
   const spotify: SpotifyWebApi.SpotifyWebApiJs = new SpotifyWebApi()
