+++ src/components/Header.tsx
@@ -1,17 +1,17 @@
-import styled from 'styled-components'
-import Button from './common/Button/Button'
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos'
 import { useHistory } from 'react-router'
+import styled from 'styled-components'
+import Button from './common/Button/Button'
 
 interface Props {
   withBack: Boolean
   children: string
 }
 
-const Header = ({ withBack, children }: Props) => {
+function Header({ withBack, children }: Props): JSX.Element {
   const history = useHistory()
 
-  const goBack = () => {
+  function goBack(): void {
     history.goBack()
   }
 
@@ -35,44 +35,44 @@ const Container = styled.div`
   position: relative;
 
   &::after {
+    background-image: linear-gradient(#5c5c5c33, transparent);
     content: '';
-    position: absolute;
-    top: 100%;
     height: 4px;
     left: 0;
+    position: absolute;
+    top: 100%;
     width: 100%;
-    background-image: linear-gradient(#5c5c5c33, transparent);
   }
 `
 const StrippedButton = styled(Button)`
-  display: flex;
   align-items: center;
-  position: absolute;
   background-color: transparent;
-  border: none;
   border-radius: 0;
-  padding: 0;
-  left: 5%;
-  top: 0;
+  border: none;
   color: var(--color-text);
-  font-size: 1.5rem;
   cursor: pointer;
+  display: flex;
+  font-size: 1.5rem;
+  left: 5%;
+  padding: 0;
+  position: absolute;
+  top: 0;
 `
 
 const StyledTitle = styled.h2`
   justify-self: center;
-  position: relative;
-  width: fit-content;
   margin-bottom: 16px;
   margin-top: 32px;
+  position: relative;
+  width: fit-content;
 
   &::after {
+    background-color: var(--color-text);
+    bottom: 0;
     content: '';
     height: 1px;
-    width: 110%;
-    position: absolute;
-    bottom: 0;
     left: -5%;
-    background-color: var(--color-text);
+    position: absolute;
+    width: 110%;
   }
 `
