+++ src/components/common/Button/Button.stories.tsx
@@ -1,4 +1,4 @@
-import { Story, Meta } from '@storybook/react'
+import { Meta, Story } from '@storybook/react'
 import Button from './Button'
 
 export default {
