+++ src/components/common/Slider/Slider.spec.tsx
@@ -1,4 +1,4 @@
-import { render, screen, fireEvent } from '@testing-library/react'
+import { fireEvent, render, screen } from '@testing-library/react'
 import Slider from './Slider'
 
 jest.mock('@material-ui/core/Slider', () => (props: any) => {
