+++ src/components/common/Slider/Slider.stories.tsx
@@ -1,4 +1,4 @@
-import { Story, Meta } from '@storybook/react'
+import { Meta, Story } from '@storybook/react'
 import Slider from './Slider'
 
 export default {
